+++ kafa/__init__.py
@@ -4,4 +4,4 @@ from __future__ import annotations
 
 __all__ = ["__version__"]
 
-__version__ = "1.24.0-beta.1"
+__version__ = "1.25.0-beta.1"